refactor: 바텀시트 컨텐츠 컴포넌트 분리
